Refactor: Use supabase-browser client for SSR and auth pages
