+++ app/auth/reset-password-otp/page.tsx
@@ -9,7 +9,7 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { ArrowLeft, Lock, Eye, EyeOff, CheckCircle, Shield } from 'lucide-react'
 import { toast } from 'sonner'
-import { supabase } from '@/lib/supabase'
+import { supabase } from '@/lib/supabase-browser'
 
 export default function ResetPasswordOtpPage() {
   const [email, setEmail] = useState('')
